// interface/products-interface.ts
export interface IProducts {
  categoryId_old?: string;
  createdAt: number;
  description: string;
  icon: string;
  id_old: string;
  image: string;
  isEstateProject: string;
  isFree: string;
  isNegotiable: string;
  maxPrice: number;
  minPrice: number;
  name: string;
  order: number;
  prefix: string;
  price: number;
  setting: any[];
  slug: string;
  status: string;
  suffixes: string;
  suggest: string;
  updatedAt: number;
  updatedBy: string;
  _id: string;
}